Rename package runner to sbt
