+++ test-sbt/jvm/src/main/scala/zio/test/sbt/ZTestRunner.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package zio.test.runner
+package zio.test.sbt
 
 import sbt.testing._
