+++ test-sbt/shared/src/main/scala/zio/test/sbt/RunnableSpecFingerprint.scala
@@ -1,4 +1,4 @@
-package zio.test.runner
+package zio.test.sbt
 
 import sbt.testing.SubclassFingerprint
 import zio.test.AbstractRunnableSpec
